+++ examples/hello_world.rs
@@ -1,6 +1,0 @@
-use bevy::prelude::*;
-use template_lib::HelloWorldPlugin;
-
-fn main() {
-    App::new().add_plugin(HelloWorldPlugin).run();
-}