+++ template_lib/src/player/mod.rs
@@ -1,0 +1,13 @@
+//! Code that connects player character logic
+use bevy::prelude::{App, Plugin};
+
+pub mod camera;
+
+/// Create a plugin to use with the main game logic
+pub struct PlayerPlugin;
+
+impl Plugin for PlayerPlugin {
+    fn build(&self, app: &mut App) {
+        app.add_plugin(camera::CameraPlugin);
+    }
+}