+++ template_lib/src/simulation/mod.rs
@@ -1,0 +1,3 @@
+/// Code to run game logic. 
+/// 
+/// This should not contain logic to render and should be able to work without a render pipeline.