+++ RELEASES.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+## Version 0.1
+
+- Update template to `bevy` 0.10
+- Improve game architecture
+- added game basics for easy use of template
+
 ## Version 0.0
 
 - Template release, plz ignore